Moved style utils inside StepperStyle
